+++ src/main/java/kernbeisser/CustomComponents/ShoppingTable/ShoppingCartController.java
@@ -14,76 +14,41 @@ public class ShoppingCartController implements Controller<ShoppingCartView,Shopp
     private ShoppingCartView view;
     private ShoppingCartModel model;
 
-    private boolean checkStorno(ShoppingItem item, boolean piece) {
-        boolean result = true;
-        boolean exit = true;
-        String response = "";
-        if (piece && item.getItemMultiplier() < 0) {
-            response = view.inputStornoRetailPrice(item.getItemRetailPrice(), false);
-            do {
-                if (response == null || response.hashCode() == 0 || response.hashCode() == 48) {
-                    exit = true;
-                    result = false;
-                } else {
-                    try {
-                        double alteredRetailPrice = Double.parseDouble(response.replace(',', '.'));
-                        if (alteredRetailPrice > 0) {
-                            if (alteredRetailPrice != item.getItemRetailPrice()) {
-                                item.setItemRetailPrice(alteredRetailPrice);
-                            }
-                            item.setName("St. " + item.getName());
-                            exit = true;
-                        } else {
-                            throw (new NumberFormatException());
-                        }
-                    } catch (NumberFormatException exception) {
-                        response = view.inputStornoRetailPrice(item.getItemRetailPrice(), true);
-                    }
-                }
-            } while (!exit);
-        } else if (!piece && item.getRetailPrice() < 0) {
-            result = (view.confirmStorno() == JOptionPane.YES_OPTION);
-        }
-        return result;
-    }
-
     public ShoppingCartController(double userValue, double userSurcharge) {
         model = new ShoppingCartModel(userValue, userSurcharge);
         view = new ShoppingCartView(this);
     }
 
     public void addShoppingItem(ShoppingItem item, boolean piece) {
-        if (checkStorno(item, piece)) {
-            int itemIndex = model.addItem(item, piece);
-            if (item.getShoppingCartIndex() == 0) {
-                item.setShoppingCartIndex(itemIndex);
-            }
-            if (item.getSingleDeposit() != 0) {
-                model.addItem(item.createItemDeposit(), true);
-            }
-            if (item.getContainerDeposit() != 0 && item.getContainerSize() > 0) {
-                if (Math.abs(item.getItemMultiplier()) >= item.getContainerSize()) {
-                    int containers = 0;
-                    boolean exit = false;
-                    String response = view.inputNoOfContainers(item, false);
-                    do {
-                        if (response == null || response.hashCode() == 0 || response.hashCode() == 48) {
-                            exit = true;
-                        } else {
-                            try {
-                                containers = Integer.parseInt(response);
-                                if (Math.signum(containers) == Math.signum(item.getItemMultiplier())) {
-                                    model.addItemBehind(item.createContainerDeposit(containers), item, true);
-                                    exit = true;
-                                } else {
-                                    throw (new NumberFormatException());
-                                }
-                            } catch (NumberFormatException exception) {
-                                response = view.inputNoOfContainers(item, true);
+        int itemIndex = model.addItem(item, piece);
+        if (item.getShoppingCartIndex() == 0) {
+            item.setShoppingCartIndex(itemIndex);
+        }
+        if (item.getSingleDeposit() != 0) {
+            model.addItem(item.createItemDeposit(), true);
+        }
+        if (item.getContainerDeposit() != 0 && item.getContainerSize() > 0) {
+            if (Math.abs(item.getItemMultiplier()) >= item.getContainerSize()) {
+                int containers = 0;
+                boolean exit = false;
+                String response = view.inputNoOfContainers(item, false);
+                do {
+                    if (response == null || response.hashCode() == 0 || response.hashCode() == 48) {
+                        exit = true;
+                    } else {
+                        try {
+                            containers = Integer.parseInt(response);
+                            if (Math.signum(containers) == Math.signum(item.getItemMultiplier())) {
+                                model.addItemBehind(item.createContainerDeposit(containers), item, true);
+                                exit = true;
+                            } else {
+                                throw (new NumberFormatException());
                             }
+                        } catch (NumberFormatException exception) {
+                            response = view.inputNoOfContainers(item, true);
                         }
-                    } while (!exit);
-                }
+                    }
+                } while (!exit);
             }
         }
         refresh();
