+++ src/main/java/kernbeisser/CustomComponents/ShoppingTable/ShoppingCartView.java
@@ -23,8 +23,6 @@ public class ShoppingCartView extends JPanel implements View<ShoppingCartControl
     private JLabel value;
     private JPanel main;
     private ObjectTable<ShoppingItem> shoppingItems;
-    private final String stornoMessageTitle = "Storno";
-
     ShoppingCartView(ShoppingCartController controller) {
         this.controller = controller;
     }
@@ -61,38 +59,7 @@ public class ShoppingCartView extends JPanel implements View<ShoppingCartControl
         }
         return response;
     }
-
-    public String inputStornoRetailPrice(double itemRetailPrice, boolean retry) {
-        String initValue = MessageFormat.format("{0, number, 0.00}", itemRetailPrice).trim();
-        String message = "";
-        String response = "";
-        if (retry) { // item is piece, first try
-            message = "Die Eingabe ist ungültig. Bitte hier einen gültigen Einzelpreis angeben, für den Fall, dass er sich seit dem ursprünglichen Einkauf geändert hat:";
-        } else { //item is piece later try
-            message = "Negative Menge: Soll der Artikel wirklich storniert werden? Dann kann hier der Einzelpreis angepasst werden, für den Fall, dass er sich seit dem ursprünglichen Einkauf geändert hat:";
-        }
-        java.awt.Toolkit.getDefaultToolkit().beep();
-        response = (String) JOptionPane.showInputDialog(
-                getContent(),
-                message,
-                stornoMessageTitle,
-                JOptionPane.YES_NO_OPTION,
-                null,
-                null,
-                initValue
-        );
-        if (response != null) {
-            response = response.trim();
-        }
-        return response;
-    }
-
-    public int confirmStorno() {
-        return JOptionPane.showConfirmDialog(
-                getContent(),"Soll die Ware wirklich storniert werden?", stornoMessageTitle, JOptionPane.YES_NO_OPTION
-        );
-    }
-
+    
     private void createUIComponents() {
         int size = 20;
         Font gridFont = new Font("Arial", Font.PLAIN, size);
