+++ src/main/java/kernbeisser/Windows/ShoppingMask/ShoppingMaskUIController.java
@@ -33,13 +33,46 @@ public class ShoppingMaskUIController implements Controller<ShoppingMaskUIView,S
         this.view = new ShoppingMaskUIView(this, shoppingCartController);
     }
 
+    private boolean checkStorno(ShoppingItem item, boolean piece) {
+        boolean result = true;
+        boolean exit = true;
+        String response = "";
+        if (piece && item.getItemMultiplier() < 0) {
+            response = view.inputStornoRetailPrice(item.getItemRetailPrice(), false);
+            do {
+                if (response == null || response.hashCode() == 0 || response.hashCode() == 48) {
+                    exit = true;
+                    result = false;
+                } else {
+                    try {
+                        double alteredRetailPrice = Double.parseDouble(response.replace(',', '.'));
+                        if (alteredRetailPrice > 0) {
+                            if (alteredRetailPrice != item.getItemRetailPrice()) {
+                                item.setItemRetailPrice(alteredRetailPrice);
+                            }
+                            item.setName("St. " + item.getName());
+                            exit = true;
+                        } else {
+                            throw (new NumberFormatException());
+                        }
+                    } catch (NumberFormatException exception) {
+                        response = view.inputStornoRetailPrice(item.getItemRetailPrice(), true);
+                    }
+                }
+            } while (!exit);
+        } else if (!piece && item.getRetailPrice() < 0) {
+            result = (view.confirmStorno() == JOptionPane.YES_OPTION);
+        }
+        return result;
+    }
+
     boolean addToShoppingCart() {
         boolean piece = (view.getOption() == ShoppingMaskUIView.ARTICLE_NUMBER || view.getOption() == ShoppingMaskUIView.CUSTOM_PRODUCT);
         boolean success = false;
         try {
             ShoppingItem item = extractShoppingItemFromUI();
-            if (item.getItemMultiplier() != 0) {
-                shoppingCartController.addShoppingItem(extractShoppingItemFromUI(), piece);
+            if (item.getItemMultiplier() != 0 && (view.getOption() == ShoppingMaskUIView.RETURN_DEPOSIT || checkStorno(item, piece) )) {
+                shoppingCartController.addShoppingItem(item, piece);
                 success = true;
             }
         } catch (UndefinedInputException undefinedInputException) {
